identity_provider/requirements.txt
bottle
pydantic
psycopg2
waitress
pyjwt